add timezones

+++ src/module/api/sendFormUsers.js
@@ -1,7 +1,7 @@
 /* eslint-disable strict */
 'use strict';
 
-import addStatus from './../addStatus.js';
+import addStatus from '../addStatus.js';
 import { authData, putUserData, deleteUserData } from './api.js';
 import creatLoader from '../creatLoader.js';
 import { getDataStorage } from '../localStorage.js';
@@ -15,11 +15,10 @@ const sendForm = (form, body) => {
 
   /* https://jsonplaceholder.typicode.com//posts/1  фейковое апи*/
   //http://192.168.99.244:9993/fsh/rest/Auth/loginSync?username=adminfsh&password=qwerty
-  //после нажатия отправить форма отчищается, а модальные окна закрываются через 3 сек
+
   if (form) {
     addStatus(form, loader.outerHTML, 60000, 'green');
     if (form.getAttribute('name') === 'authorization') {
-      console.log('authorization');
       authData(body.URL, body)
         .then(response => {
           if (response.status !== 200) {
@@ -32,6 +31,7 @@ const sendForm = (form, body) => {
     }
 
     if (form.getAttribute('name') === 'row') {
+      addStatus(form, loader.outerHTML, 60000, 'green');
       const URL = getDataStorage('URL');
       const getData = getDataStorage('name');
       putUserData(URL, body, getData.tokenValue)
@@ -43,19 +43,32 @@ const sendForm = (form, body) => {
         })
         .then(resultRow.bind(this, form, body))
         .catch(error.bind(this, form));
+      /*if (body.file.name !== '') {
+        loadPhotoMultipart(URL, getData.tokenValue, { 'file': body.file })
+         .then(response => {
+          if (response.status !== 200) {
+            throw 'Unable connect to server URL address !!!';
+          }
+          return (response.text());
+        })
+        .then(resultRow.bind(this, form, body))
+        .catch(error.bind(this, form)); 
+      }*/
     }
-  } else {
-    const URL = getDataStorage('URL');
-    const getData = getDataStorage('name');
-    deleteUserData(URL, getData.tokenValue, body)
-      .then(response => {
-        if (response.status !== 200) {
-          throw 'Unable connect to server URL address !!!';
-        }
-        return (response.text());
-      })
-      .then(updateData.bind(this, form))
-      .catch(error.bind(this, form));
+    if (form.getAttribute('name') === 'delete') {
+      addStatus(form, loader.outerHTML, 60000, 'green');
+      const URL = getDataStorage('URL');
+      const getData = getDataStorage('name');
+      deleteUserData(URL, getData.tokenValue, body)
+        .then(response => {
+          if (response.status !== 200) {
+            throw 'Unable connect to server URL address !!!';
+          }
+          return (response.text());
+        })
+        .then(updateData.bind(this, form))
+        .catch(error.bind(this, form));
+    }
   }
 };
